requirements.txt
asgiref==3.7.2
astor==0.8.1
attrs==23.2.0
bandit==1.7.6
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
darglint==1.8.1
Django==5.0.1
djangorestframework==3.14.0
docutils==0.20.1
entrypoints==0.4
eradicate==2.3.0
flake8==7.0.0
flake8-bandit==4.1.1
flake8-broken-line==1.0.0
flake8-bugbear==23.12.2
flake8-commas==2.1.0
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-isort==6.1.1
flake8-quotes==3.3.2
flake8-rst-docstrings==0.3.0
flake8-string-format==0.3.0
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
pbr==6.0.0
pep8-naming==0.13.3
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.17.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
restructuredtext-lint==1.4.0
rich==13.7.0
smmap==5.0.1
snowballstemmer==2.2.0
sqlparse==0.4.4
stevedore==5.1.0
toml==0.10.2
typing_extensions==4.9.0
urllib3==2.1.0
wemake-python-styleguide==0.18.0
